step down the master if the kill file is present
if --kill-file-should-trigger-step-down is passed to the process and the
process is the master of a replica set, then automatically step down
when the kill file is observed.

also, force the node to marked as unelectable as long as the kill file
is present.

also rename the fileExists method to isKilled since it's clearer