remove touchfile and healthbinary modules
We're going to rely on the kill-file, and we can have an external binary
that's long-lived and decides when to create it, including decisions
about how many failures should trigger the kill-file, what the criteria
are, etc.

The kill-file and the binary allow the same level of customization and
control in the end.